Make generated headers static for easier diff
With this commit, generated Java files no longer contain the generated
date and time and the generator version in the header as default. This
makes comparison between different generated versions with diff tools
easier. The previous default changed generated files on every generator
run, even if it the only change was the date and time of the last
generator run.

Signed-off-by: Dietmar Stoll <eclipse.etc@dietmar-stoll.de>